Convert benchmark tests into pytester tests
